Still running into an issue
Compiling error, where function sendStats has a void parameter (illegal), and the parameter is also not declared in this scope. As you can see, everything is commented out, so this shouldn't be an issue. It seems to be compiling using a _Communication file, which was the old name before I changed it to Comms. ????